Reorganize anchors management
